Allow loading configuration from environment
For things like Heroku deployment where application.yml makes life difficult
