+++ app/mailers/mailer.rb
@@ -1,6 +1,6 @@
 class Mailer < ActionMailer::Base
-  default :from => Juvia::Application.config.from
-  uri = URI.parse(Juvia::Application.config.base_url)
+  default :from => (ENV['JUVIA_EMAIL_FROM'] || Juvia::Application.config.from)
+  uri = URI.parse(ENV['JUVIA_BASE_URL'] || Juvia::Application.config.base_url)
   default_url_options[:protocol] = uri.scheme
   default_url_options[:host] = uri.host
