+++ config/initializers/secret_token.rb
@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Juvia::Application.config.secret_token = '7c09a74e53a97a2ebf5328ce5333e6634502931587ce75d2fefc36aa5b9427d71f31645bd4e6001118f0673aec991caec9f61645538ec6fe4e6992c4aef501f8'
+Juvia::Application.config.secret_token = ENV['JUVIA_SESSION_SECRET'] || '7c09a74e53a97a2ebf5328ce5333e6634502931587ce75d2fefc36aa5b9427d71f31645bd4e6001118f0673aec991caec9f61645538ec6fe4e6992c4aef501f8'
